Add repr rule to op
